Improved the "Button" UI component (#61)
