+++ src/components/ui/Button.tsx
@@ -1,0 +1,45 @@
+import { cn } from '@/lib/tailwind/utils';
+
+import { Slot } from '@radix-ui/react-slot';
+import { cva, type VariantProps } from 'class-variance-authority';
+import * as React from 'react';
+
+const buttonVariants = cva(
+  'text-base min-w-[70px] rounded inline-flex items-center justify-center gap-2.5',
+  {
+    variants: {
+      variant: {
+        primary: 'bg-primary text-white hover:bg-primary-dark',
+        error: 'bg-danger text-white hover:bg-danger-dark',
+        cancel: 'bg-transparent text-black hover:bg-gray-light',
+      },
+      size: {
+        sm: 'h-8 px-7',
+        md: 'h-10 px-12',
+        lg: 'h-12 px-12',
+      },
+    },
+    defaultVariants: {
+      variant: 'primary',
+      size: 'md',
+    },
+  },
+);
+
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
+    VariantProps<typeof buttonVariants> {
+  asChild?: boolean;
+}
+
+const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
+  ({ className, variant, size, asChild = false, ...props }, ref) => {
+    const Comp = asChild ? Slot : 'button';
+    return (
+      <Comp className={cn(buttonVariants({ variant, size, className }))} ref={ref} {...props} />
+    );
+  },
+);
+Button.displayName = 'Button';
+
+export { Button, buttonVariants };
