+++ src/components/ui/index.ts
@@ -1,3 +1,3 @@
 export { BottomTabIconLink } from './BottomTabIconLink';
-export { Button, buttonVariants } from './button';
+export { Button, buttonVariants } from './Button';
 export { Icon } from './Icon';
